Gets detailed information of books from Open Library Books
In order to facilitate the work of the librarians, the client
requested that the detailed information of the books must be
obtained through the Open Library Books platform.
Since this platform provides an end point to obtain the detailed
information of a book by providing its ISBN, it was requested to
store only the titles of the books and the ISBNs of the catalog in
the database.

The tables of the database that stored the information of the
authors are eliminated. In addition, a column is added to the books
table to add the ISBN. Along with this, related database querie
are updated.
Also, the `sugbi.catalog.open-library-books` namespace that handles
requests to the Open Library Books endpoint is created. And
finally, a namespace `sugbi.catalog.core` is created to mix the
information from the database with  Open Library Books.