Adjusts function parameters of `delete-book!`
For consistency reasons, it is recommended that records in the
database be distinguished by the use of an identifier. In the books
table, we have two identifiers: `isbn`, and `book_id`. The latter
was added for internal use of the database, so users will use
`isbn` to refer to the books.
At first the `delete-book!` function was implemented using the
`title` of the book as a parameter, however, as already mentioned,
this is not convenient. For example, a book registered as
"How to Build a Car`, could generate errors when users do not use
the correct set of capital letters, complicating the interface.

Adjusted the HugSQL function definition `delete-books!` to take
`isbn` instead of `title` as a parameter.