fixup! Use kebab-case in the Open Library Books database and queries
It was forgotted to adjust the names of the fields that are
obtained from the external Open Library Books service. When
changing from snake_case to kebab-case, the keys did not match, so
the information is not retrieved correctly.

Adjusted the names of the `relevant-fields` in the
`sugbi.catalog.open-library-books` namespace to use kebab-case.