Adds routes to operate with books
Since a REST API will be used to allow users to interact with
information on the server, it is necessary to start building the
routes that will be able to access. In previous commits, the
functions to interact with the database and the external Open
Library Books service have already been built.

Added GET, POST and DELETE methods of the `api/catalog/books` route
to allow searching, adding and deleting books respectively, using
previously defined functions.