Refactors catalog end points
Initially the first version of the end points was implemented in
the `sugbi.routes.services` namespace. This file is intended to add
all the paths of the operations that will be available on the
system. However, since there will be several entities, it is not
convenient to define the handlers directly on it, cause the file
will grow quite quickly. Similarly, handlers were implemented
"manually". Defining the response maps for each request. Details of
these can be removed using the functions of the `ring` library.

- `sugbi.catalog.handlers` and `sugbi.catalog.routes` namespaces
  are created to have a better code distribution.
- The functions of the `ring.util.http-response` namespace are used
  to create the responses of the end points. Using this library
  simplifies the code and allows to maintain a natural language
  instead of directly specifying HTTP tags.
- Relocated the `catalog.db/delete-book!` endpoint to adhere to
  Vinay Sahni's recommendations about RESTful APIs.
- In the same sense, the end point of
  `catalog.core/enriched-search-books-by-title` was adjusted to
  obtain the complete catalog in case a search criteria is not
  passed as a parameter.
- Created the `relevant-fields` variable in the
  `sugbi.catalog.open-library-books` namespace to allow quick
  access to the available book fields.
- Finally, the search by title was adjusted in order to not
  distinguish upper and lower case letters, simplifying searches.