Adds authentication with Google
To avoid implementing user management manually, Google will be used
as a authentication service. This is because the system will be
used by members of the Faculty who already have Google accounts. In
addition, to allowing development focused in the context of the
problem instead of focusing efforts on implementing an
authentication module that should be agile and secure. Google uses
the OAuth 2.0 and OpenID Connect protocols so it can be considered
secure enough.

Added the `auth.google.core` namespace that implements all the
requests of the OAuth 2.0 protocol to keep the details of it
encapsulated in a single module. Hence, handlers defined in the
`auth.google.handlers` namespace can make use of this protocol at a
high level.
In addition to keeping the `client_secret` secure, the
`dev-config.edn` file was modified to add the variables needed by
Google to authorize Sugbi to authenticate users. This file should
be kept within `.gitignore` but is kept in the repository for
teaching purposes.