Adds the librarians to the model
Since the user authentication flow is already implemented, now it
is necessary to be able to distinguish those who are librarians.
These roles cannot be defined in Google, as they pertain
exclusively to the application context. So they must be defined
inside Sugbi.
OAuth 2.0 defines a value called `sub` which identifies each user.
This value will be taken as it is unique and permanent. A user's
email or name can change over time, but this value will always be
maintained to identify the user.

Migrations are added to create a `user_management.librarian` table
that stores the `sub` of the users who are librarians. In addition
to this, the `sugbi.user-management.db` and
`sugbi.user-management.core` namespaces are created to add
functions that operate with this new entity.
Finally, the `callback` of the authentication with Google was
adjusted to add to the session an `is-librarian?` variable that
determines if a user is a librarian or not.