Restricts access to librarian operations
Some operations, such as adding or removing books, should only be
possible for librarians.

The book add and delete handlers were adjusted to verify that the
user has an active session and that that session belongs to a
librarian. If not, the operation is not executed and an HTTP 403
code is returned.