Validates spec on catalog end points
To ensure that we send the correct information from our REST API,
the schema of the data that is requested and transmitted on each
end point must be added.

Adds the catalog specs and configure the routes to use them.
Since `data-spec` does not recognize sequences as a collection, the
`sugbi.catalog.core/merge-on-key` function is adjusted to return a
vector instead of a sequence.