Extends the Reagent sample component
To enable the BMI slider, the `calc-bmi` function is modified to
calculate a value given the input map keys. If it contains weight
and height, it will calculate the bmi. If it contains height and
bmi, it will calculate weight.
It also adds a small label in the interface to show the diagnosis.