Adds support for book cover images
It has been requested that the system allow storing images of some
book covers. These images may be in JPG or SVG format, and must be
stored in the database.

The database model is adjusted to add a couple of columns where the
image data will be stored as well as its file extension. In
addition, the 'end point' for book registration is adjusted so that
from now, a cover image must be provided. Finally, a new route is
added to be able to consult the image of the book if it exists.

The `byte-streams` library was added to be able to serialize the
information to the necessary formats in each system layer without
directly using Java objects.